Splited Batch to Writer and Batch.
